In founding vote confirmation, explain the three-day voting period.
Some people found it confusing that the three-day voting period for the
founding vote was different from the voting period they had chosen in their
constitution for ordinary proposals. Add some copy to the confirmation dialog
to explain that the three-day founding vote period is fixed, and separate from
the voting period in their constitution.