Short-circuit STI explicitly instead of using abstract_class.
Classes like FoundingMember (a descendent of Member) exist only to
allow convenient syntax like FoundingMember.new. It's not intended
for the 'type' column' to ever contain 'FoundingMember'; it should
just have 'Member.

Previously, we achieved this by setting 'abstract_class' true for
these convenience subclasses. This causes the class to be ignored
by Rails' STI machinery when calculating what should go in the
'type' column.

In Rails 4, abstract classes are not allowed to be instantiated.
So we have stopped making these abstract classes, and instead
tell Active Record explicitly what to put in the 'type' column for
these classes, by using the 'sti_name' method.